pyproject.toml
[tool.poetry]
name = "file2text"
version = "0.1.0"
description = "Python package to read pdf / word file and extract text from it."
authors = ["Beth Ou Yang"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
